move crossover function
